refactor: split up COSE openssl module into multiple files
